Quick hack to fix pod names for addons
This addresses the issue when multiple copies of `ember-component-css` are running in the build (e.g. when developing using linked/local addons/engines which have their own `node_modules` folders), the copies end up overwriting each other's `pod-names.js` file (and the logic for reading existing values fails cause the separate plugins operate in different temp directories).

The end result was that the CSS was being generated correctly, but the `componentCssClassName` was failing to find the pod name for any components other than those in the last package/addon that `ember-component-css` ran for (e.g. the last one to write `pod-names.js`).

Using `global` is a dirty hack, but I don't understand broccoli well enough to know how to correctly remedy this problem yet.  With more research eventually a fix which prevents multiple plugins being created, or allows them to reliably share a single `pod-names.js`, would be the proper fix (which may be pushable upstream then); for now though, this will suffice as something which works (albeit not in an ideal fashion).